indexer cleanup 3/N: lib, db and indexer.rs
