+++ .github/actions/ts-e2e/action.yml
@@ -65,7 +65,7 @@ runs:
 
     - name: Set env
       run: |
-        echo "E2E_RUN_LOCAL_NET_CMD=(RUST_LOG=\"consensus=off\" $(echo $PWD/target/debug/sui-test-validator) --with-indexer --use-indexer-v2 --pg-port 5432 --pg-db-name sui_indexer_v2 --graphql-host 127.0.0.1 --graphql-port 9125)" >> $GITHUB_ENV
+        echo "E2E_RUN_LOCAL_NET_CMD=(RUST_LOG=\"consensus=off\" $(echo $PWD/target/debug/sui-test-validator) --with-indexer --pg-port 5432 --pg-db-name sui_indexer_v2 --graphql-host 127.0.0.1 --graphql-port 9125)" >> $GITHUB_ENV
         echo "VITE_SUI_BIN=$PWD/target/debug/sui" >> $GITHUB_ENV
       shell: bash
