+++ crates/sui-cluster-test/src/config.rs
@@ -31,16 +31,10 @@ pub struct ClusterTestOpt {
     /// URL for the indexer RPC server
     #[clap(long)]
     pub indexer_address: Option<String>,
-    /// Use new version of indexer or not
-    #[clap(long)]
-    pub use_indexer_v2: bool,
     /// URL for the Indexer Postgres DB
     #[clap(long)]
     #[derivative(Debug(format_with = "obfuscated_pg_address"))]
     pub pg_address: Option<String>,
-    /// TODO(gegao): remove this after indexer migration is complete.
-    #[clap(long)]
-    pub use_indexer_experimental_methods: bool,
     #[clap(long)]
     pub config_dir: Option<PathBuf>,
     /// URL for the indexer RPC server
@@ -72,10 +66,8 @@ impl ClusterTestOpt {
             epoch_duration_ms: None,
             indexer_address: None,
             pg_address: None,
-            use_indexer_experimental_methods: false,
             config_dir: None,
             graphql_address: None,
-            use_indexer_v2: false,
         }
     }
 }
