+++ crates/sui-indexer/src/lib.rs
@@ -3,30 +3,34 @@
 #![recursion_limit = "256"]
 
 use std::net::SocketAddr;
-use std::{collections::HashMap, time::Duration};
 
 use anyhow::{anyhow, Result};
-use axum::{extract::Extension, http::StatusCode, routing::get, Router};
 use clap::Parser;
-use diesel::pg::PgConnection;
-use diesel::r2d2::ConnectionManager;
 use jsonrpsee::http_client::{HeaderMap, HeaderValue, HttpClient, HttpClientBuilder};
 use metrics::IndexerMetrics;
-use prometheus::{Registry, TextEncoder};
-use regex::Regex;
-use tracing::{info, warn};
+use prometheus::Registry;
+use tokio::runtime::Handle;
+use tracing::warn;
 use url::Url;
 
-use errors::IndexerError;
-use mysten_metrics::RegistryService;
+use sui_json_rpc::ServerType;
+use sui_json_rpc::{JsonRpcServerBuilder, ServerHandle};
 use sui_json_rpc_api::CLIENT_SDK_TYPE_HEADER;
 
+use crate::apis::{
+    CoinReadApiV2, ExtendedApiV2, GovernanceReadApiV2, IndexerApiV2, MoveUtilsApiV2, ReadApiV2,
+    TransactionBuilderApiV2, WriteApi,
+};
+use crate::indexer_reader::IndexerReader;
+use errors::IndexerError;
+
 pub mod apis;
+pub mod db;
 pub mod errors;
 pub mod framework;
 mod handlers;
+pub mod indexer;
 pub mod indexer_reader;
-pub mod indexer_v2;
 pub mod metrics;
 pub mod models_v2;
 pub mod processors_v2;
@@ -34,28 +38,6 @@ pub mod schema;
 pub mod store;
 pub mod test_utils;
 pub mod types;
-pub mod utils;
-
-pub type PgConnectionPool = diesel::r2d2::Pool<ConnectionManager<PgConnection>>;
-pub type PgPoolConnection = diesel::r2d2::PooledConnection<ConnectionManager<PgConnection>>;
-
-const METRICS_ROUTE: &str = "/metrics";
-/// Returns all endpoints for which we have implemented on the indexer,
-/// some of them are not validated yet.
-/// NOTE: we only use this for integration testing
-const IMPLEMENTED_METHODS: [&str; 9] = [
-    // read apis
-    "get_checkpoint",
-    "get_latest_checkpoint_sequence_number",
-    "get_object",
-    "get_owned_objects",
-    "get_total_transaction_blocks",
-    "get_transaction_block",
-    "multi_get_transaction_blocks",
-    // indexer apis
-    "query_events",
-    "query_transaction_blocks",
-];
 
 #[derive(Parser, Clone, Debug)]
 #[clap(
@@ -86,8 +68,6 @@ pub struct IndexerConfig {
     pub rpc_server_url: String,
     #[clap(long, default_value = "9000", global = true)]
     pub rpc_server_port: u16,
-    #[clap(long, num_args(1..))]
-    pub migrated_methods: Vec<String>,
     #[clap(long)]
     pub reset_db: bool,
     #[clap(long)]
@@ -96,11 +76,6 @@ pub struct IndexerConfig {
     pub rpc_server_worker: bool,
     #[clap(long)]
     pub analytical_worker: bool,
-    // NOTE: experimental only, do not use in production.
-    #[clap(long)]
-    pub skip_db_commit: bool,
-    #[clap(long)]
-    pub use_v2: bool,
 }
 
 impl IndexerConfig {
@@ -116,10 +91,6 @@ impl IndexerConfig {
             url.host_str().unwrap_or_default(),
             url.port().unwrap_or_default()
         ))
-    }
-
-    pub fn all_implemented_methods() -> Vec<String> {
-        IMPLEMENTED_METHODS.iter().map(|&s| s.to_string()).collect()
     }
 
     pub fn get_db_url(&self) -> Result<String, anyhow::Error> {
@@ -150,15 +121,40 @@ impl Default for IndexerConfig {
             client_metric_port: 9184,
             rpc_server_url: "0.0.0.0".to_string(),
             rpc_server_port: 9000,
-            migrated_methods: vec![],
             reset_db: false,
             fullnode_sync_worker: true,
             rpc_server_worker: true,
             analytical_worker: false,
-            skip_db_commit: false,
-            use_v2: false,
         }
     }
+}
+
+pub async fn build_json_rpc_server(
+    prometheus_registry: &Registry,
+    reader: IndexerReader,
+    config: &IndexerConfig,
+    custom_runtime: Option<Handle>,
+) -> Result<ServerHandle, IndexerError> {
+    let mut builder = JsonRpcServerBuilder::new(env!("CARGO_PKG_VERSION"), prometheus_registry);
+    let http_client = crate::get_http_client(config.rpc_client_url.as_str())?;
+
+    builder.register_module(WriteApi::new(http_client.clone()))?;
+    builder.register_module(IndexerApiV2::new(reader.clone()))?;
+    builder.register_module(TransactionBuilderApiV2::new(reader.clone()))?;
+    builder.register_module(MoveUtilsApiV2::new(reader.clone()))?;
+    builder.register_module(GovernanceReadApiV2::new(reader.clone()))?;
+    builder.register_module(ReadApiV2::new(reader.clone()))?;
+    builder.register_module(CoinReadApiV2::new(reader.clone()))?;
+    builder.register_module(ExtendedApiV2::new(reader.clone()))?;
+
+    let default_socket_addr: SocketAddr = SocketAddr::new(
+        // unwrap() here is safe b/c the address is a static config.
+        config.rpc_server_url.as_str().parse().unwrap(),
+        config.rpc_server_port,
+    );
+    Ok(builder
+        .start(default_socket_addr, custom_runtime, Some(ServerType::Http))
+        .await?)
 }
 
 fn get_http_client(rpc_client_url: &str) -> Result<HttpClient, IndexerError> {
@@ -177,166 +173,4 @@ fn get_http_client(rpc_client_url: &str) -> Result<HttpClient, IndexerError> {
                 e
             ))
         })
-}
-
-pub fn new_pg_connection_pool(
-    db_url: &str,
-    pool_size: Option<u32>,
-) -> Result<PgConnectionPool, IndexerError> {
-    let pool_config = PgConnectionPoolConfig::default();
-    let manager = ConnectionManager::<PgConnection>::new(db_url);
-
-    let pool_size = pool_size.unwrap_or(pool_config.pool_size);
-    diesel::r2d2::Pool::builder()
-        .max_size(pool_size)
-        .connection_timeout(pool_config.connection_timeout)
-        .connection_customizer(Box::new(pool_config.connection_config()))
-        .build(manager)
-        .map_err(|e| {
-            IndexerError::PgConnectionPoolInitError(format!(
-                "Failed to initialize connection pool with error: {:?}",
-                e
-            ))
-        })
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct PgConnectionPoolConfig {
-    pool_size: u32,
-    connection_timeout: Duration,
-    statement_timeout: Duration,
-}
-
-impl PgConnectionPoolConfig {
-    const DEFAULT_POOL_SIZE: u32 = 100;
-    const DEFAULT_CONNECTION_TIMEOUT: u64 = 30;
-    const DEFAULT_STATEMENT_TIMEOUT: u64 = 30;
-
-    fn connection_config(&self) -> PgConnectionConfig {
-        PgConnectionConfig {
-            statement_timeout: self.statement_timeout,
-            read_only: false,
-        }
-    }
-
-    pub fn set_pool_size(&mut self, size: u32) {
-        self.pool_size = size;
-    }
-
-    pub fn set_connection_timeout(&mut self, timeout: Duration) {
-        self.connection_timeout = timeout;
-    }
-
-    pub fn set_statement_timeout(&mut self, timeout: Duration) {
-        self.statement_timeout = timeout;
-    }
-}
-
-impl Default for PgConnectionPoolConfig {
-    fn default() -> Self {
-        let db_pool_size = std::env::var("DB_POOL_SIZE")
-            .ok()
-            .and_then(|s| s.parse::<u32>().ok())
-            .unwrap_or(Self::DEFAULT_POOL_SIZE);
-        let conn_timeout_secs = std::env::var("DB_CONNECTION_TIMEOUT")
-            .ok()
-            .and_then(|s| s.parse::<u64>().ok())
-            .unwrap_or(Self::DEFAULT_CONNECTION_TIMEOUT);
-        let statement_timeout_secs = std::env::var("DB_STATEMENT_TIMEOUT")
-            .ok()
-            .and_then(|s| s.parse::<u64>().ok())
-            .unwrap_or(Self::DEFAULT_STATEMENT_TIMEOUT);
-
-        Self {
-            pool_size: db_pool_size,
-            connection_timeout: Duration::from_secs(conn_timeout_secs),
-            statement_timeout: Duration::from_secs(statement_timeout_secs),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-struct PgConnectionConfig {
-    statement_timeout: Duration,
-    read_only: bool,
-}
-
-impl diesel::r2d2::CustomizeConnection<PgConnection, diesel::r2d2::Error> for PgConnectionConfig {
-    fn on_acquire(&self, conn: &mut PgConnection) -> std::result::Result<(), diesel::r2d2::Error> {
-        use diesel::{sql_query, RunQueryDsl};
-
-        sql_query(format!(
-            "SET statement_timeout = {}",
-            self.statement_timeout.as_millis(),
-        ))
-        .execute(conn)
-        .map_err(diesel::r2d2::Error::QueryError)?;
-
-        if self.read_only {
-            sql_query("SET default_transaction_read_only = 't'")
-                .execute(conn)
-                .map_err(diesel::r2d2::Error::QueryError)?;
-        }
-
-        Ok(())
-    }
-}
-
-pub fn get_pg_pool_connection(pool: &PgConnectionPool) -> Result<PgPoolConnection, IndexerError> {
-    pool.get().map_err(|e| {
-        IndexerError::PgPoolConnectionError(format!(
-            "Failed to get connection from PG connection pool with error: {:?}",
-            e
-        ))
-    })
-}
-
-fn convert_url(url_str: &str) -> Option<String> {
-    // NOTE: unwrap here is safe because the regex is a constant.
-    let re = Regex::new(r"https?://([a-z0-9-]+\.[a-z0-9-]+\.[a-z]+)").unwrap();
-    let captures = re.captures(url_str)?;
-
-    captures.get(1).map(|m| m.as_str().to_string())
-}
-
-pub fn start_prometheus_server(
-    addr: SocketAddr,
-    fn_url: &str,
-) -> Result<(RegistryService, Registry), anyhow::Error> {
-    let converted_fn_url = convert_url(fn_url);
-    if converted_fn_url.is_none() {
-        warn!(
-            "Failed to convert full node url {} to a shorter version",
-            fn_url
-        );
-    }
-    let fn_url_str = converted_fn_url.unwrap_or_else(|| "unknown_url".to_string());
-
-    let labels = HashMap::from([("indexer_fullnode".to_string(), fn_url_str)]);
-    info!("Starting prometheus server with labels: {:?}", labels);
-    let registry = Registry::new_custom(Some("indexer".to_string()), Some(labels))?;
-    let registry_service = RegistryService::new(registry.clone());
-
-    let app = Router::new()
-        .route(METRICS_ROUTE, get(metrics))
-        .layer(Extension(registry_service.clone()));
-
-    tokio::spawn(async move {
-        axum::Server::bind(&addr)
-            .serve(app.into_make_service())
-            .await
-            .unwrap();
-    });
-    Ok((registry_service, registry))
-}
-
-async fn metrics(Extension(registry_service): Extension<RegistryService>) -> (StatusCode, String) {
-    let metrics_families = registry_service.gather_all();
-    match TextEncoder.encode_to_string(&metrics_families) {
-        Ok(metrics) => (StatusCode::OK, metrics),
-        Err(error) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            format!("unable to encode metrics: {error}"),
-        ),
-    }
 }
