+++ crates/sui-indexer/src/store/module_resolver_v2.rs
@@ -15,6 +15,7 @@ use sui_types::base_types::{ObjectID, SequenceNumber};
 use sui_types::move_package::MovePackage;
 use sui_types::object::Object;
 
+use crate::db::PgConnectionPool;
 use crate::errors::{Context, IndexerError};
 use crate::handlers::tx_processor::IndexingPackageBuffer;
 use crate::metrics::IndexerMetrics;
@@ -22,7 +23,6 @@ use crate::models_v2::packages::StoredPackage;
 use crate::schema::{objects, packages};
 use crate::store::diesel_macro::read_only_blocking;
 use crate::types::IndexedPackage;
-use crate::PgConnectionPool;
 
 /// A package resolver that reads packages from the database.
 pub struct IndexerStorePackageModuleResolver {
